+++ src/main/java/bowling/domain/score/scores/DefaultFrameScores.java
@@ -1,4 +1,4 @@
-package bowling.domain.score;
+package bowling.domain.score.scores;
 
 public class DefaultFrameScores extends Scores {
 }
