+++ src/main/java/bowling/domain/score/scores/LastFrameScores.java
@@ -1,4 +1,4 @@
-package bowling.domain.score;
+package bowling.domain.score.scores;
 
 public class LastFrameScores extends Scores {
 }
